BinGenerationTests: checkpoint.
fixed a number of bugs, but the result is still not correct.
I know that this commit also contains debug messages, will be removed later.

Also added a number of data files and directories required to execute the code
correctly. In the C# version they are donwloaded at run time, but I would like to avoid that in the C++ version.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>