BinGenerationTest: some cleanup
not every outputfile is correct, but I am turning to another test at this point, and this commit
removes unnecessary printf statements.

The final files that are not yet correct all depend on the ProteinQuantificationEngine in FlashLFQ.
As far as I can see, the input values to the FlashLFQ engine are correct (i.e. identical to the C# version),
the result out of the FlashFLQ engine is however different.

After discussion with the MetaMorpheusTeam it is however clear, that a) this part of FlashLFQ went through an update earlier this year, and b) this part is not relevant for XLsearch.

Therefore, I am postponing debugging this part for now, and focus on other tests. Will return at a later stage.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>