XLTest.cpp: make the DeadendTest compile and run.
Disabling for now writing the indices using cereal, since it is a) a bit buggy and b) very slow.
It is in fact right now faster to recalculate the indices every time.

The XL_Intralinks.tsv file generated as part of hte DeadendTest is mostly correct, but there are a few items that I will have to go over. THis is to be done next.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>