XLSearchOutputTest.cpp: make test compile and run
the output is mostly correct, although it complains about 45 vs. 44 tab separated entries. Probably a miossing
or miss-identifiued double tab that is not parsed correctly.

The two values that differ from the C# version are the rank again (is 40 in the C# version but 68 in the C++ version),
and the Beta Peptide Rank ( 28 in C#, 56 in C++ )

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>